Remove and ignore Cargo.lock
The content of Cargo.lock varies based on which OS it was generated on, and might also vary based on which feature flags are enabled.
The project doesn't have a reference OS and have multiple possible combinations of feature flags.
For that reason having a Cargo.lock doesn't make sense.