Improve negation-edge release logic, and other improvements
Changes RELEASE message to include the deepest component that may be released. Only a single component should be release at a time.
Changes TOKEN message to include the deepest component depth seen by any worker. This is used to know which depth can be released safely.
Removes counter based release logic. This have been superseded by the token ring algorithm.
Modifies Worker::release_negations to take the depth of the deepest component it may release.
Changes Worker::have_pending_negation_edges to also account for negation edges in Worker::unsafe_neg_edges.
Adds a sanity check to Worker::release_negations that check if more that one component is being released.
Removes worker ID from some tracing events because it's already contained in the span.
Adds a sanity check to Worker::final_assign to check that an assignment is not downgraded or changed from TRUE to FALSE.
Adds a comment to edg.rs::test::test_1 that illustrates the graph used in the test.